Safari-609.1.11.1/LayoutTests/fast/events/before-input-prevent-cut-expected.txt

To manually test this, select the text in the input and attempt to cut. The value of the input should not change.

initial value:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "helloworld"

after prevented cut:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "helloworld"

after allowed cut:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       <br>
